Crash fixes (was dereferencing empty lists).
svn path=/trunk/KDE/kdeedu/kmplot/; revision=521720

+++ kmplot/kconstanteditor.cpp
@@ -89,7 +89,7 @@ void KConstantEditor::cmdDelete_clicked()
         for( Q3ValueVector<Ufkt>::iterator it =  m_view->parser()->ufkt.begin(); it !=  m_view->parser()->ufkt.end(); ++it)
 	{
 		str =  it->fstr;
-		for (int i=str.find(')'); (uint)i<str.length();i++)
+		for (int i=str.find(')'); i<str.length();i++)
 			if ( str.at(i) == constant )
                         {
 			     KMessageBox::error(this, i18n("A function uses this constant; therefore, it cannot be removed."));
